Fix bug with deleted subscriptions due to duplicated entries in the ConnectionFlag table. Delete subscriptions only if a PushNotificationException is raised and not if there was something else unexpected raised.

+++ foss_public_alert_server/subscriptionHandler/tasks.py
@@ -61,10 +61,13 @@ class NotificationBaseTask(Task):
         :param einfo:
         :return: None
         """
-        # get the subscription and delete it
-        subscription_id = args[0]
-        logger.debug(f"delete subscription {subscription_id}")
-        Subscription.objects.get(id=subscription_id).delete()
+        # Only delete the subscriptions if we raised a PushNotificatioinException.
+        # This avoids deleting subscriptions due to internal errors
+        if isinstance(exc, PushNotificationException):
+            # get the subscription and delete it
+            subscription_id = args[0]
+            logger.debug(f"delete subscription {subscription_id}")
+            Subscription.objects.get(id=subscription_id).delete()
 
 @shared_task(name="task.send_notification",
              bind=True,
